ObjectMasonry View
