+++ package.json
@@ -36,7 +36,6 @@
     "vue-i18n": "^9.2.2",
     "vue-json-csv": "^2.1.0",
     "vue-router": "^4.1.5",
-    "@yeger/vue-masonry-wall": "^3.4.4",
     "vue-flex-waterfall": "^2.2.2"
   },
   "devDependencies": {
