+++ src/components/core/CoreUsersDetail.vue
@@ -1,17 +1,16 @@
 <script setup lang="ts">
-import ObjectDetailNew from "@/components/datatable/ObjectDetailNew.vue";
+import ObjectDetail from "@/components/datatable/ObjectDetail.vue";
+import ObjectMasonry from "@/components/datatable/ObjectMasonry.vue";
 import { useMembersStore } from "@/stores/members";
 import { useMenuStore } from "@/stores/menu";
 import type { endpoints } from "@/api/api";
-import { useI18n } from "vue-i18n";
-import { VueFlexWaterfall } from "vue-flex-waterfall";
-const { t } = useI18n();
+
 const menuStore = useMenuStore();
 const membersStore = useMembersStore();
 menuStore.setTitle("User");
 // TODO: Load username from store
 // TODO: Add memberships and tags
-// TODO: Make profile list dynamic
+// TODO: Make profile list dynamic (based on extensions)
 const profile_endpoints: Array<keyof typeof endpoints> = [
   "coreUsers",
   "profilesProfiles",
@@ -21,32 +20,10 @@ const profile_endpoints: Array<keyof typeof endpoints> = [
 </script>
 
 <template>
-  <VueFlexWaterfall
-    align-content="stretch"
-    class="w-full"
-    col="3"
-    col-spacing="18px"
-    :break-at="{ 1536: 2, 768: 1 }"
-  >
-    <div v-for="item in profile_endpoints" class="waterfall-item mb-5">
-      <ObjectDetailNew
-        :store="membersStore"
-        :name="item"
-        :id="$route.params.id"
-      >
-      </ObjectDetailNew>
-    </div>
-  </VueFlexWaterfall>
+  <ObjectMasonry :items="profile_endpoints">
+    <template #item="{ item }">
+      <ObjectDetail :store="membersStore" :name="item" :id="$route.params.id">
+      </ObjectDetail>
+    </template>
+  </ObjectMasonry>
 </template>
-
-<style lang="scss">
-.waterfall-item {
-  width: 100%;
-  @media screen and (min-width: 768px) {
-    width: calc(50% - 9px);
-  }
-  @media screen and (min-width: 1536px) {
-    width: calc(33.333% - 6px);
-  }
-}
-</style>
